Warbler: completed test cases

+++ warbler/app.py
@@ -347,8 +347,11 @@ def messages_destroy(message_id):
         return redirect("/")
 
     msg = Message.query.get(message_id)
-    db.session.delete(msg)
-    db.session.commit()
+    if msg.user == g.user:
+        db.session.delete(msg)
+        db.session.commit()
+    else:
+        flash("Access unauthorized.", "danger")
 
     return redirect(f"/users/{g.user.id}")
